+++ tests/Fixtures/SomeMixin.php
@@ -18,4 +18,14 @@ class SomeMixin
             return Str::lower($value);
         };
     }
+
+    public function havingVariadic()
+    {
+        return function(string &... $params): void {
+            $params[] = 'foo';
+            foreach ($params as $param) {
+                echo $param;
+            }
+        };
+    }
 }